refactor(front): dissolve components/ into app/layout and shared/ui
- app chrome -> app/layout: BarraBrasil, MenuBar (only used by RootLayout)
- reusable UI -> shared/ui: MapSvg, MapBrazil, DownloadButtons, filtroDados,
  SortableTable, ResponsiveChartContainer, Toast/
- removed the duplicate components/SimaTable re-export shim (canonical lives
  in shared/ui)
- updated all importers to @/ alias paths
- typed SortableTable rows (Record<string,unknown>) removing an explicit any
