--- front/src/components/SimaTable.tsx
@@ -1,2 +1,0 @@
-/** Reexport: implementação em `shared/ui` (Frente 2 SIMA). */
-export { default } from "../shared/ui/SimaTable";
