+++ front/src/features/sima/tables/hooks/useSimaTablesPage.ts
@@ -1,6 +1,6 @@
 import { useCallback, useEffect, useMemo, useReducer, useRef, useState } from "react";
-import MapBrazil from "@/components/MapBrazil";
-import { useToast } from "@/components/Toast/useToast";
+import MapBrazil from "@/shared/ui/MapBrazil";
+import { useToast } from "@/shared/ui/Toast/useToast";
 import { messageForFailedSimaExport } from "../api/simaExportFeedback";
 import { simaDownloadPath, type SimaTableId } from "../../api/simaEndpoints";
 import {
