+++ front/src/shared/ui/MapSvg.tsx
