+++ front/src/shared/ui/ResponsiveChartContainer.tsx
